requirements.txt
fastapi[all]
python-multipart
python-jose
passlib
bcrypt
sqlalchemy
PyJWT
psycopg2
python-dotenv
asyncpg
alembic
gunicorn
paramiko
pandas